just want to know

+++ common/src/stack/command/stack/commands/remove/firmware/make/plugin_basic.py
@@ -14,46 +14,47 @@ import stack.commands
 from stack.util import unique_everseen, lowered
 from stack.exception import ArgRequired, ArgError, CommandError
 
+
 class Plugin(stack.commands.Plugin):
-	"""Attempts to remove all provided makes and any associated models and firmware from the database."""
+    """Attempts to remove all provided makes and any associated models and firmware from the database."""
 
-	def provides(self):
-		return "basic"
+    def provides(self):
+        return "basic"
 
-	def remove_related_models(self, makes):
-		"""Remove any models related to the provided makes."""
-		# remove associated models
-		for make in makes:
-			# get all the models associated with this make
-			models_to_remove = [
-				row[0] for row in
-				self.owner.db.select(
-					"""
+    def remove_related_models(self, makes):
+        """Remove any models related to the provided makes."""
+        # remove associated models
+        for make in makes:
+            # get all the models associated with this make
+            models_to_remove = [
+                row[0]
+                for row in self.owner.db.select(
+                    """
 					firmware_model.name
 					FROM firmware_model
 						INNER JOIN firmware_make
 							ON firmware_model.make_id=firmware_make.id
 					WHERE firmware_make.name=%s
 					""",
-					make,
-				)
-			]
+                    make,
+                )
+            ]
 
-			# and remove them if we found any
-			if models_to_remove:
-				self.owner.call(command = "remove.firmware.model", args = [*models_to_remove, f"make={make}"])
+            # and remove them if we found any
+            if models_to_remove:
+                self.owner.call(
+                    command="remove.firmware.model",
+                    args=[*models_to_remove, f"make={make}"],
+                )
 
-	def run(self, args):
-		# get rid of any duplicate names
-		makes = tuple(unique_everseen(lowered(args)))
-		# ensure the make names already exist
-		self.owner.ensure_makes_exist(makes = makes)
+    def run(self, args):
+        # get rid of any duplicate names
+        makes = tuple(unique_everseen(lowered(args)))
+        # ensure the make names already exist
+        self.owner.ensure_makes_exist(makes=makes)
 
-		# Remove any related models.
-		self.remove_related_models(makes = makes)
+        # Remove any related models.
+        self.remove_related_models(makes=makes)
 
-		# now delete the makes
-		self.owner.db.execute(
-			"DELETE FROM firmware_make WHERE name IN %s",
-			(makes,)
-		)
+        # now delete the makes
+        self.owner.db.execute("DELETE FROM firmware_make WHERE name IN %s", (makes,))
